backend/requirements.txt
Flask==3.1.1
Flask-CORS==5.0.0
pymongo==4.10.1
redis==5.0.8
python-dotenv==1.0.1
requests==2.32.3
Werkzeug==3.1.3
flask-jwt-extended==4.6.0
APScheduler==3.10.4
requests==2.32.3

